// src/app/records/patient/model/patient.ts
export interface Patient {
  patientId?: number;
  identificationId?: number;
  identificationNumber?: string;
  fileNumber?: number;
  inssNumber?: string;
  maritalStatusId?: number;
  email?: string;
  genderId?: number;
  nationalityId?: number;
  countryBornInId?: number;
  departmentBornInId?: number;
  municipalityBornInId?: number;
  countryResidenceId?: number;
  departmentResidenceId?: number;
  municipalityResidenceId?: number;
  bloodTypeId?: number;
  professionId?: number;
  religionId?: number;
  firstName?: string;
  secondName?: string;
  firstSurname?: string;
  secondSurname?: string;
  dateOfBirth?: Date;
  homeAddress?: string;
  homePhone?: string;
  mobilePhone?: string;
  active?: string;
  pregnant?: string;
  deceased?: string;
  status?: number;
}